REFACTOR : konempty PR Review 반영
소스코드 파일에 마지막 개행 추가, setFragment의 매개변수를 String이 아닌 FragmentList enum을 받도록 수정, 사용하지 않는 문자열 리소스 제거, View와 변수를 연결하는 방식을 Kotlin Synthetic에서 View Binding으로 변경
