+++ .idea/misc.xml
@@ -5,7 +5,7 @@
       <map>
         <entry key="..\:/Users/qnfpd/fakedevelopers/app/src/main/res/layout/activity_main.xml" value="0.2875" />
         <entry key="..\:/Users/qnfpd/fakedevelopers/app/src/main/res/layout/fragment_login.xml" value="0.25" />
-        <entry key="..\:/Users/qnfpd/fakedevelopers/app/src/main/res/layout/fragment_login_type.xml" value="0.25" />
+        <entry key="..\:/Users/qnfpd/fakedevelopers/app/src/main/res/layout/fragment_login_type.xml" value="0.1" />
       </map>
     </option>
   </component>
