+++ app/src/androidTest/java/com/fakedevelopers/ddangddangmarket/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.fakedevelopers.ddangddangmarket_aos
+package com.fakedevelopers.ddangddangmarket
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
