+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.fakedevelopers.ddangddangmarket_aos">
+    package="com.fakedevelopers.ddangddangmarket">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.DDangDDangMarket_AOS">
         <activity
-            android:name=".MainActivity"
+            android:name="com.fakedevelopers.ddangddangmarket.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
