+++ app/src/main/java/com/fakedevelopers/ddangddangmarket/FragmentList.kt
@@ -1,0 +1,5 @@
+package com.fakedevelopers.ddangddangmarket
+
+enum class FragmentList {
+    LOGIN, LOGINTYPE
+}
