+++ app/src/main/res/layout/fragment_login.xml
@@ -47,7 +47,6 @@
         android:layout_marginEnd="24dp"
         android:autofillHints=""
         android:hint="@string/login_password"
-        android:imeActionLabel="@string/action_sign_in_short"
         android:imeOptions="actionDone"
         android:inputType="textPassword"
         android:selectAllOnFocus="true"
