+++ app/src/main/res/layout/fragment_login_type.xml
@@ -32,8 +32,7 @@
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/imageView_logintype_logo"
-        app:layout_constraintVertical_bias="0.05
-" />
+        app:layout_constraintVertical_bias="0.05" />
 
     <Button
         android:id="@+id/button_logintype_kakaologin"
