+++ app/src/test/java/com/fakedevelopers/ddangddangmarket/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.fakedevelopers.ddangddangmarket_aos
+package com.fakedevelopers.ddangddangmarket
 
 import org.junit.Test
